refactor: move pencil icon to common
